requirements.txt
Flask_Login==0.6.3
Werkzeug==3.0.0
Flask==3.0.0
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.25
email-validator
Pillow
WTForms==3.0.1
gunicorn==23.0.0
psycopg2-binary>=2.9.9
python-dotenv==1.0.1